backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
mysqlclient==2.2.6
reportlab==4.2.5
